bugfix
- wrong arguments were passed in OD_checker to queue, caused the OD_checker to be disfunctional.

- MySQL connector was imperfectly implemented, caused the commands to be out of sync

+++ DBmanager/measurement.py
@@ -47,7 +47,7 @@ class PeriodicalMeasurement(Thread):
                 data = json.dumps(data)
                 post('https://localhost:5000/' + str(self.node_id) + '/' + str(device), data=data, headers=headers, verify=False, auth=('BioArInEO', 'sybila'))
 
-            sleep(int(self.experiment_details.get('sleep_time', 60)))
+            sleep(float(self.experiment_details.get('sleep_time', 60)))
